gitignore
Quitado target y classpath

+++ src/main/java/org/eu5/ainhoalm/airportAena/dao/AirplaneDAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.eu5.ainhoalm.airportAena.model.Airplane;
 
-public interface AirplaneDAO extends GenericDAO<Airplane, Integer> {
+public interface AirplaneDAO extends GenericDAO<Airplane, Long> {
 	public abstract Airplane findByKey(String registration);
 	public abstract List<Airplane> findByState(int code);
 }
