+++ src/main/java/org/eu5/ainhoalm/airportAena/dao/AirplaneStateDAO.java
@@ -2,6 +2,6 @@ package org.eu5.ainhoalm.airportAena.dao;
 
 import org.eu5.ainhoalm.airportAena.model.AirplaneState;
 
-public interface AirplaneStateDAO extends GenericDAO<AirplaneState, Integer> {
+public interface AirplaneStateDAO extends GenericDAO<AirplaneState, Long> {
 	public abstract AirplaneState findByKey(int code);
 }
