+++ src/main/java/org/eu5/ainhoalm/airportAena/dao/impl/hibernate/AirplaneStateDAOHibernateImpl.java
@@ -8,7 +8,7 @@ import org.eu5.ainhoalm.airportAena.utils.HibernateUtil;
 import org.hibernate.Query;
 import org.hibernate.Session;
 
-public class AirplaneStateDAOHibernateImpl extends GenericDAOHibernateImpl<AirplaneState, Integer> implements AirplaneStateDAO {
+public class AirplaneStateDAOHibernateImpl extends GenericDAOHibernateImpl<AirplaneState, Long> implements AirplaneStateDAO {
 
 	@SuppressWarnings("unchecked")
 	@Override
